+++ src/java/org/malbino/orion/facades/GrupoFacade.java
@@ -12,6 +12,7 @@ import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 import org.malbino.orion.entities.Grupo;
+import org.malbino.orion.entities.Mencion;
 import org.malbino.orion.enums.Nivel;
 import org.malbino.orion.enums.Turno;
 
@@ -50,7 +51,7 @@ public class GrupoFacade extends AbstractFacade<Grupo> {
 
         return l;
     }
-    
+
     public List<Grupo> listaGrupos(int id_gestionacademica, int id_carrera, Nivel nivel) {
         List<Grupo> l = new ArrayList();
 
@@ -202,7 +203,7 @@ public class GrupoFacade extends AbstractFacade<Grupo> {
 
         return l;
     }
-    
+
     public List<String> listaParalelos(int id_gestionacademica, int id_carrera, Nivel nivel) {
         List<String> l = new ArrayList();
 
@@ -219,7 +220,7 @@ public class GrupoFacade extends AbstractFacade<Grupo> {
 
         return l;
     }
-    
+
     public List<String> listaParalelos(int id_gestionacademica, int id_carrera, Nivel nivel, Turno turno) {
         List<String> l = new ArrayList();
 
@@ -227,6 +228,44 @@ public class GrupoFacade extends AbstractFacade<Grupo> {
             Query q = em.createQuery("SELECT DISTINCT g.codigo FROM Grupo g JOIN g.gestionAcademica ga JOIN g.materia m JOIN m.carrera c WHERE ga.id_gestionacademica=:id_gestionacademica AND c.id_carrera=:id_carrera AND m.nivel=:nivel AND g.turno=:turno ORDER BY g.codigo");
             q.setParameter("id_gestionacademica", id_gestionacademica);
             q.setParameter("id_carrera", id_carrera);
+            q.setParameter("nivel", nivel);
+            q.setParameter("turno", turno);
+
+            l = q.getResultList();
+        } catch (Exception e) {
+
+        }
+
+        return l;
+    }
+
+    public List<String> listaParalelos(int id_gestionacademica, int id_carrera, Mencion mencion, Nivel nivel, Turno turno) {
+        List<String> l = new ArrayList();
+
+        try {
+            Query q = em.createQuery("SELECT DISTINCT g.codigo FROM Grupo g JOIN g.gestionAcademica ga JOIN g.materia m JOIN m.carrera c WHERE ga.id_gestionacademica=:id_gestionacademica AND c.id_carrera=:id_carrera AND (m.mencion IS NULL OR m.mencion=:mencion) AND m.nivel=:nivel AND g.turno=:turno ORDER BY g.codigo");
+            q.setParameter("id_gestionacademica", id_gestionacademica);
+            q.setParameter("id_carrera", id_carrera);
+            q.setParameter("mencion", mencion);
+            q.setParameter("nivel", nivel);
+            q.setParameter("turno", turno);
+
+            l = q.getResultList();
+        } catch (Exception e) {
+
+        }
+
+        return l;
+    }
+    
+    public List<Grupo> listaGrupos(int id_gestionacademica, int id_carrera, Mencion mencion, Nivel nivel, Turno turno) {
+        List<Grupo> l = new ArrayList();
+
+        try {
+            Query q = em.createQuery("SELECT g FROM Grupo g JOIN g.gestionAcademica ga JOIN g.materia m JOIN m.carrera c WHERE ga.id_gestionacademica=:id_gestionacademica AND c.id_carrera=:id_carrera AND (m.mencion IS NULL OR m.mencion=:mencion) AND m.nivel=:nivel AND g.turno=:turno ORDER BY m.nombre, g.codigo");
+            q.setParameter("id_gestionacademica", id_gestionacademica);
+            q.setParameter("id_carrera", id_carrera);
+            q.setParameter("mencion", mencion);
             q.setParameter("nivel", nivel);
             q.setParameter("turno", turno);