+++ src/java/org/malbino/orion/servlets/LibroInscripcionesCarrera.java
@@ -28,27 +28,26 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import org.malbino.orion.entities.Carrera;
+import org.malbino.orion.entities.GestionAcademica;
 import org.malbino.orion.entities.Inscrito;
-import org.malbino.orion.entities.Instituto;
-import org.malbino.orion.enums.Nivel;
-import org.malbino.orion.enums.Periodo;
+import org.malbino.orion.facades.CarreraFacade;
+import org.malbino.orion.facades.GestionAcademicaFacade;
 import org.malbino.orion.facades.InscritoFacade;
-import org.malbino.orion.facades.InstitutoFacade;
-import org.malbino.orion.util.Constantes;
 
 /**
  *
  * @author tincho
  */
-@WebServlet(name = "LibroInscripciones", urlPatterns = {"/reportes/LibroInscripciones"})
-public class LibroInscripciones extends HttpServlet {
+@WebServlet(name = "LibroInscripcionesCarrera", urlPatterns = {"/reportes/LibroInscripcionesCarrera"})
+public class LibroInscripcionesCarrera extends HttpServlet {
 
     private static final String CONTENIDO_PDF = "application/pdf";
 
-    private static final Font TITULO = FontFactory.getFont(FontFactory.HELVETICA, 10, Font.BOLD, BaseColor.BLACK);
-    private static final Font SUBTITULO = FontFactory.getFont(FontFactory.HELVETICA, 8, Font.NORMAL, BaseColor.BLACK);
-    private static final Font NEGRITA = FontFactory.getFont(FontFactory.HELVETICA, 6.5f, Font.BOLD, BaseColor.BLACK);
-    private static final Font NORMAL = FontFactory.getFont(FontFactory.HELVETICA, 6.5f, Font.NORMAL, BaseColor.BLACK);
+    private static final Font TITULO = FontFactory.getFont(FontFactory.HELVETICA, 12, Font.BOLD, BaseColor.BLACK);
+    private static final Font SUBTITULO = FontFactory.getFont(FontFactory.HELVETICA, 10, Font.NORMAL, BaseColor.BLACK);
+    private static final Font NEGRITA = FontFactory.getFont(FontFactory.HELVETICA, 8, Font.BOLD, BaseColor.BLACK);
+    private static final Font NORMAL = FontFactory.getFont(FontFactory.HELVETICA, 8, Font.NORMAL, BaseColor.BLACK);
 
     private static final int MARGEN_IZQUIERDO = -40;
     private static final int MARGEN_DERECHO = -40;
@@ -56,7 +55,9 @@ public class LibroInscripciones extends HttpServlet {
     private static final int MARGEN_INFERIOR = 30;
 
     @EJB
-    InstitutoFacade institutoFacade;
+    GestionAcademicaFacade gestionAcademicaFacade;
+    @EJB
+    CarreraFacade carreraFacade;
     @EJB
     InscritoFacade inscritoFacade;
 
@@ -71,10 +72,12 @@ public class LibroInscripciones extends HttpServlet {
     }
 
     public void generarPDF(HttpServletRequest request, HttpServletResponse response) {
-        Integer gestion = (Integer) request.getSession().getAttribute("gestion");
-        Periodo periodo = (Periodo) request.getSession().getAttribute("periodo");
+        Integer id_gestionacademica = (Integer) request.getSession().getAttribute("id_gestionacademica");
+        Integer id_carrera = (Integer) request.getSession().getAttribute("id_carrera");
 
-        if (gestion != null && periodo != null) {
+        if (id_gestionacademica != null && id_carrera != null) {
+            GestionAcademica gestionAcademica = gestionAcademicaFacade.find(id_gestionacademica);
+            Carrera carrera = carreraFacade.find(id_carrera);
             try {
                 response.setContentType(CONTENIDO_PDF);
 
@@ -83,23 +86,22 @@ public class LibroInscripciones extends HttpServlet {
 
                 document.open();
 
-                document.add(titulo(gestion, periodo));
-                document.add(contenido(gestion, periodo));
+                document.add(titulo(gestionAcademica, carrera));
+                document.add(contenido(gestionAcademica, carrera));
 
                 document.close();
             } catch (IOException | DocumentException ex) {
-                Logger.getLogger(LibroInscripciones.class.getName()).log(Level.SEVERE, null, ex);
+                Logger.getLogger(LibroInscripcionesCarrera.class.getName()).log(Level.SEVERE, null, ex);
             }
 
         }
     }
 
-    public PdfPTable titulo(Integer gestion, Periodo periodo) throws BadElementException, IOException {
+    public PdfPTable titulo(GestionAcademica gestionAcademica, Carrera carrera) throws BadElementException, IOException {
         PdfPTable table = new PdfPTable(100);
 
         //cabecera
-        Instituto instituto = institutoFacade.buscarPorId(Constantes.ID_INSTITUTO);
-        String realPath = getServletContext().getRealPath("/resources/uploads/" + instituto.getLogo());
+        String realPath = getServletContext().getRealPath("/resources/uploads/" + carrera.getCampus().getInstituto().getLogo());
         Image image = Image.getInstance(realPath);
         image.scaleToFit(70, 70);
         image.setAlignment(Image.ALIGN_CENTER);
@@ -111,26 +113,26 @@ public class LibroInscripciones extends HttpServlet {
         cell.setBorder(Rectangle.NO_BORDER);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase("LIBRO DE INSCRIPCIONES,", TITULO));
+        cell = new PdfPCell(new Phrase("LIBRO DE INSCRIPCIONES POR CARRERA,", TITULO));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(80);
         cell.setBorder(Rectangle.NO_BORDER);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
-
-        cell = new PdfPCell(new Phrase(instituto.getNombreRegulador(), SUBTITULO));
+        
+        cell = new PdfPCell(new Phrase(carrera.getCampus().getInstituto().getNombreRegulador(), SUBTITULO));
+        cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
+        cell.setColspan(80);
+        cell.setBorder(Rectangle.NO_BORDER);
+        table.addCell(cell);
+        
+        cell = new PdfPCell(new Phrase(gestionAcademica.toString(), SUBTITULO));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(80);
         cell.setBorder(Rectangle.NO_BORDER);
         table.addCell(cell);
 
-        cell = new PdfPCell(new Phrase(String.valueOf(gestion), SUBTITULO));
-        cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
-        cell.setColspan(80);
-        cell.setBorder(Rectangle.NO_BORDER);
-        table.addCell(cell);
-
-        cell = new PdfPCell(new Phrase(periodo.getPeriodoRomano(), SUBTITULO));
+        cell = new PdfPCell(new Phrase(carrera.toString(), SUBTITULO));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
         cell.setColspan(80);
         cell.setBorder(Rectangle.NO_BORDER);
@@ -139,12 +141,12 @@ public class LibroInscripciones extends HttpServlet {
         return table;
     }
 
-    public PdfPTable contenido(Integer gestion, Periodo periodo) throws BadElementException, IOException {
-        PdfPTable table = new PdfPTable(140);
+    public PdfPTable contenido(GestionAcademica gestionAcademica, Carrera carrera) throws BadElementException, IOException {
+        PdfPTable table = new PdfPTable(100);
 
         PdfPCell cell = new PdfPCell(new Phrase(" ", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
-        cell.setColspan(140);
+        cell.setColspan(100);
         cell.setBorder(Rectangle.NO_BORDER);
         table.addCell(cell);
 
@@ -156,37 +158,25 @@ public class LibroInscripciones extends HttpServlet {
 
         cell = new PdfPCell(new Phrase("Codigo", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-        cell.setColspan(13);
+        cell.setColspan(10);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
         cell = new PdfPCell(new Phrase("Estudiante", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-        cell.setColspan(30);
-        cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
-        table.addCell(cell);
-
-        cell = new PdfPCell(new Phrase("Carrera", NEGRITA));
-        cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-        cell.setColspan(30);
-        cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
-        table.addCell(cell);
-
-        cell = new PdfPCell(new Phrase("Nivel", NEGRITA));
-        cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-        cell.setColspan(10);
+        cell.setColspan(35);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
         cell = new PdfPCell(new Phrase("Tipo", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-        cell.setColspan(11);
+        cell.setColspan(10);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
         cell = new PdfPCell(new Phrase("Fecha", NEGRITA));
         cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-        cell.setColspan(11);
+        cell.setColspan(10);
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
@@ -196,7 +186,7 @@ public class LibroInscripciones extends HttpServlet {
         cell.setBackgroundColor(BaseColor.LIGHT_GRAY);
         table.addCell(cell);
 
-        List<Inscrito> listaInscritos = inscritoFacade.listaInscritos(gestion, periodo);
+        List<Inscrito> listaInscritos = inscritoFacade.listaInscritos(gestionAcademica.getId_gestionacademica(), carrera.getId_carrera());
         for (int i = 0; i < listaInscritos.size(); i++) {
             Inscrito inscrito = listaInscritos.get(i);
 
@@ -209,49 +199,32 @@ public class LibroInscripciones extends HttpServlet {
             cell = new PdfPCell(new Phrase(String.valueOf(inscrito.getCodigo()), NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
-            cell.setColspan(13);
+            cell.setColspan(10);
             table.addCell(cell);
 
             cell = new PdfPCell(new Phrase(inscrito.getEstudiante().toString(), NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
-            cell.setColspan(30);
-            table.addCell(cell);
-
-            cell = new PdfPCell(new Phrase(inscrito.getCarrera().toString(), NORMAL));
-            cell.setHorizontalAlignment(PdfPCell.ALIGN_LEFT);
-            cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
-            cell.setColspan(30);
-            table.addCell(cell);
-
-            Nivel nivel = inscrito.nivel();
-            if (nivel != null) {
-                cell = new PdfPCell(new Phrase(String.valueOf(inscrito.nivel().getAbreviatura()), NORMAL));
-            } else {
-                cell = new PdfPCell(new Phrase("", NORMAL));
-            }
-            cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
-            cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
-            cell.setColspan(10);
+            cell.setColspan(35);
             table.addCell(cell);
 
             cell = new PdfPCell(new Phrase(inscrito.getTipo().toString(), NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
-            cell.setColspan(11);
+            cell.setColspan(10);
             table.addCell(cell);
 
             cell = new PdfPCell(new Phrase(inscrito.fecha_ddMMyyyy(), NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
-            cell.setColspan(11);
+            cell.setColspan(10);
             table.addCell(cell);
 
             cell = new PdfPCell(new Phrase(" ", NORMAL));
             cell.setHorizontalAlignment(PdfPCell.ALIGN_CENTER);
             cell.setVerticalAlignment(PdfPCell.ALIGN_MIDDLE);
             cell.setColspan(30);
-            cell.setFixedHeight(20);
+            cell.setFixedHeight(30);
             table.addCell(cell);
         }