+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2023-03-11
+### Added
+- Libro de Inscripciones por Carrera
+### Changed
+- Libro de Inscripciones
+- Planilla de Seguimiento
+### Removed
+### Fixed
+- Seguimiento Academico
+
 ## [2.0.6] - 2023-03-07
 ### Added
 ### Changed